Add func to fetch all languages used in a repository
When the data of a repository are fetched, the JSON object
doesn't include the details of all languages used, instead,
it only specifies the main language for this repository.

This commit add func to fetch the list of languages used and their
LOCs. However, the names of the languages are not fetched as a list,
they act as the attributes of the JSON object, which makes it hard to
deserialize them directly by RestClient. This JSON object for the
languages are fetched as a JSON String first and desiralised after
using custom deserializer.